Add new question block
This is a simple alternative to the ProblemBlock,
with similar functionality, but a reduced featureset.
Only features currently used on LabXChange are implemented.

It is fully compatible with the olx from a ProblemBlock,
for the features it does support.
Any unrecognised olx is ignored.

The student view isn't used,
because we upload olx directly to edit the question
(with the assessment editor),
and use the custom student_view_user_state handler
to provide the json data for rendering the block
for students on the frontend.

Also make pytest more verbose for tests to help with debugging.